Updated pre-commit config too.
